improve presentation users in the dialogues list
For dialogues between users where neither is the current user,
the user without mod/dialogue:manage permission is presented
as the primary in the dialogues list. Also, when viewing a dialogue
between two users neither of whom is the current user, the 'Dialogue
between X and Y' form of message is presented.